refactor - move logic from mainNavigator to Main
