+++ App/App.js
@@ -4,11 +4,14 @@ import { registerRootComponent } from 'expo';
 
 import store from "store";
 import AppTheme from "theme/AppTheme";
+import RootNavigation from "navigation/RootNavigation";
 
 const App = () => {
     return (
         <Provider store={store}>
-            <AppTheme />
+            <AppTheme>
+							<RootNavigation />
+						</AppTheme>
         </Provider>
     );
 };
