+++ App/theme/AppTheme.js
@@ -3,19 +3,13 @@ import { useSelector } from "react-redux";
 import { ThemeProvider } from "emotion-theming";
 
 import themes from "theme/themes";
-import RootNavigation from "navigation/RootNavigation";
 import { selectors } from "store";
 
-const AppTheme = () => {
+const AppTheme = ({ children }) => {
     const themeType = useSelector(selectors.theme);
     const selectedTheme = themes[themeType];
     
-
-    return (
-        <ThemeProvider theme={selectedTheme}>
-            <RootNavigation />
-        </ThemeProvider>
-    );
+    return <ThemeProvider theme={selectedTheme}>{children}</ThemeProvider>;
 };
 
 export default AppTheme;
